Extend benchmark
